# sample-project/snapshot/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Store API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/orders/admin/all": {
      "get": {
        "operationId": "orders.adminListAllOrders",
        "summary": "GET /api/v1/orders/admin/all",
        "tags": ["orders"],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminListAllOrdersOrdersResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {}
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {}
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdminListAllOrdersOrdersResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "result": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "totalPrice": {
                  "type": "number"
                },
                "status": {
                  "type": "string"
                },
                "userId": {
                  "type": "string"
                }
              },
              "required": ["id", "totalPrice", "status", "userId"]
            }
          }
        },
        "required": ["success", "result"]
      },
...
    }
  }
}